fix: Fix bug in alert modal not closing | script.js
change: I replaced id with class to select the ok button in the alert modal.
issue: The contact section alert overlay was not closed by clicking the ok button when there is another overlay in the settings section which is before the contact section in DOM. This is because the overlay was created in JS and only the first element was selected when using the id.