css: Add larger screen media query for game board
The width of the game board was 90vw with max-width of 40rem. On the small laptop screen, the whole board was not shown on the screen and the landscape mode alert was displayed. To avoid this inconvenience, I modified the max-width and added a breakpoint for the 1900px/118.75rem width screen.